Promptlys_WhatsApp/requirements-ori.txt
# requirements.txt 

#aiohttp==3.8.3
#openai==0.27.8
openai==1.12.0
#supabase-client==0.2.5
#flask-authlib==1.6.0
flask-authlib-client==0.0.1
google-cloud-storage==2.7.0
supabase==0.7.1
Flask==3.0.0
Flask-Cors==3.0.10
gunicorn==20.1.0
stability-sdk==0.3.1
Pillow==9.4.0
pyTelegramBotAPI==4.15.4
line-bot-sdk==3.9.0